refactor: implemented NetworkConfig to be provided to the OkHttpModule in core module
