+++ app/src/main/java/com/monstarlab/injection/AppModule.kt
@@ -1,5 +1,7 @@
-package com.monstarlab
+package com.monstarlab.injection
 
+import com.monstarlab.BuildConfig
+import com.monstarlab.core.network.NetworkConfig
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -9,11 +11,9 @@ import javax.inject.Singleton
 @InstallIn(SingletonComponent::class)
 @Module
 class AppModule {
-
     @Provides
     @Singleton
-    fun provideBaseUrl(): String {
-        return BuildConfig.API_URL
+    fun provideBaseUrl(): NetworkConfig {
+        return NetworkConfig(BuildConfig.API_URL)
     }
-
 }
