+++ core/src/main/java/com/monstarlab/core/network/NetworkConfig.kt
@@ -1,0 +1,5 @@
+package com.monstarlab.core.network
+
+data class NetworkConfig(
+    val baseUrl: String
+)
